Clustering Big Data
This file contains the code to use homogeneity analysis to determine an optimal representation of the categorical attributes in a large data set. The data here is the airline delay data for 2016. Stratified sampling is used to obtain a sample such that all origin destination pairs are represented. Homogeneity Analysis is then run on this sample to learn an optimal representation. The learned representation is used to recode the original data to a numeric representation. This file is then clustered with mini batch K means.